+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* [iOS] Added iOS [FWDebug](https://github.com/lszzy/FWDebug) library for Flutter.
+* [iOS] Added [Inspector](https://github.com/kekland/inspector) library for Flutter.